+++ packages/orchestrator/cmd/mount-build-rootfs/main.go
@@ -15,6 +15,7 @@ import (
 
 	"github.com/e2b-dev/infra/packages/orchestrator/cmd/internal/cmdutil"
 	"github.com/e2b-dev/infra/packages/orchestrator/pkg/sandbox/block"
+	"github.com/e2b-dev/infra/packages/orchestrator/pkg/sandbox/nbd/nbdutil"
 	"github.com/e2b-dev/infra/packages/orchestrator/pkg/sandbox/nbd/testutils"
 	"github.com/e2b-dev/infra/packages/shared/pkg/featureflags"
 	"github.com/e2b-dev/infra/packages/shared/pkg/logger"
@@ -116,7 +117,7 @@ func runEmpty(ctx, nbdContext context.Context, featureFlags *featureflags.Client
 	overlay := block.NewOverlay(emptyDevice, cache)
 	defer overlay.Close()
 
-	devicePath, deviceCleanup, err := testutils.GetNBDDevice(nbdContext, testutils.NewLoggerOverlay(overlay), featureFlags)
+	devicePath, deviceCleanup, err := nbdutil.GetNBDDevice(nbdContext, testutils.NewLoggerOverlay(overlay), featureFlags)
 	defer deviceCleanup.Run(ctx, 30*time.Second)
 	if err != nil {
 		return fmt.Errorf("failed to get nbd device: %w", err)
@@ -132,7 +133,7 @@ func runEmpty(ctx, nbdContext context.Context, featureFlags *featureflags.Client
 }
 
 func run(ctx, nbdContext context.Context, featureFlags *featureflags.Client, buildID, mountPath string, verify bool) error {
-	rootfs, rootfsCleanup, err := testutils.TemplateRootfs(ctx, buildID)
+	rootfs, rootfsCleanup, err := nbdutil.TemplateRootfs(ctx, buildID)
 	defer rootfsCleanup.Run(ctx, 30*time.Second)
 	if err != nil {
 		return fmt.Errorf("failed to get template rootfs: %w", err)
@@ -157,7 +158,7 @@ func run(ctx, nbdContext context.Context, featureFlags *featureflags.Client, bui
 	overlay := block.NewOverlay(rootfs, cache)
 	defer overlay.Close()
 
-	devicePath, deviceCleanup, err := testutils.GetNBDDevice(nbdContext, overlay, featureFlags)
+	devicePath, deviceCleanup, err := nbdutil.GetNBDDevice(nbdContext, overlay, featureFlags)
 	defer deviceCleanup.Run(ctx, 30*time.Second)
 	if err != nil {
 		return fmt.Errorf("failed to get nbd device: %w", err)
@@ -173,7 +174,7 @@ func run(ctx, nbdContext context.Context, featureFlags *featureflags.Client, bui
 
 		fmt.Fprintf(os.Stdout, "creating mount path directory: %s\n", mountPath)
 
-		mountCleanup, err := testutils.MountNBDDevice(devicePath, mountPath)
+		mountCleanup, err := nbdutil.MountNBDDevice(devicePath, mountPath)
 		defer mountCleanup.Run(ctx, 30*time.Second)
 		if err != nil {
 			return fmt.Errorf("failed to mount device to mount path: %w", err)
